feat: Enhance OpenRouter integration with new properties and services
- Added configuration properties for OpenRouter API including connection settings, timeouts, and default model.
- Implemented OpenRouterProperties class to manage OpenRouter configurations.
- Created ServiceException class for better error handling in service layer.
- Developed ChatCompletionService to handle chat completion requests and responses.
- Introduced PromptEnhancementService for improving user prompts before sending to AI.
- Implemented StreamingService for handling streaming responses from OpenRouter.
- Created OpenRouterClient to manage API calls with resilience features like circuit breaker and retry.
- Added LoggingUtils for structured logging and request context management.
- Developed RequestBodyBuilder to construct API request bodies from AIRequest objects.
- Implemented ResponseExtractor to parse and extract relevant data from API responses.
- Enhanced application.yml with new settings for resilience and rate limiting.

+++ llm-service/src/main/java/com/craftpilot/llmservice/service/ChatService.java
@@ -40,14 +40,14 @@ public class ChatService {
                 .map(aiResponse -> mapToChatResponse(aiResponse, chatRequest));
     }
 
-    private ChatResponse mapToChatResponse(AIResponse aiResponse, ChatRequest request) {
-        Integer tokensUsed = aiResponse.getTokensUsed();
+    private ChatResponse mapToChatResponse(AIResponse aiResponse, ChatRequest request) { 
+        long tokensUsed = aiResponse.getTokensUsed();
         return ChatResponse.builder()
                 .userId(request.getUserId())
                 .message(request.getMessage())
                 .response(aiResponse.getResponse())
                 .model(aiResponse.getModel())
-                .processedTokens(tokensUsed)
+                .processedTokens((int)tokensUsed) // long değeri int'e cast ediyoruz
                 .timestamp(System.currentTimeMillis())
                 .build();
     }
